fix(gate): устранить предупреждения clippy и выпустить 0.7.1
Ворота релизного процесса гоняют clippy с флагом -D warnings, где предупреждение
становится ошибкой и останавливает сборку до job build, поэтому артефакты релиза
v0.7.0 не собрались вовсе. Устранены шесть предупреждений: лишние приведения типа к
самому себе в answers.rs и model.rs, лишний вызов trim перед split_whitespace в
engines/cli.rs, избыточный ok() при сопоставлении с Some в model.rs, бесполезный
format! без подстановок в interview.rs, а также утверждение о константах в тесте
runner.rs, заменённое на строго более сильную проверку времени компиляции.

Версия рабочей области поднята с 0.7.0 до 0.7.1: выпуск 0.7.0 не дал артефактов из-за
падения ворот, а тег не перезаписывается на месте, поэтому исправление выходит новой
версией. Снимок публичного контракта и документы по государственным стандартам
пересобраны под 0.7.1, расхождений с кодом нет.

Проверено дословным прогоном всех трёх ворот: cargo fmt --check без расхождений,
cargo clippy --workspace --all-targets -- -D warnings без единого предупреждения,
cargo test --workspace без падений, плюс контрольная release-сборка бинаря (31 МБ).

+++ crates/ailc-core/src/answers.rs
@@ -144,7 +144,7 @@ pub fn today() -> String {
 fn civil_from_days(z: i64) -> (i64, u32, u32) {
     let z = z + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64;
+    let doe = z - era * 146_097;
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
